#50 Introduced better folder structure to the project.
