+++ package.json
@@ -8,44 +8,44 @@
   "source": "src/main.ts",
   "types": "src/main.ts",
   "exports": {
-    ".": {
-      "import": "./lib/main.esm.js",
-      "require": "./lib/main.umd.js",
-      "default": "./lib/main.umd.js"
-    }
+	".": {
+	  "import": "./lib/main.esm.js",
+	  "require": "./lib/main.umd.js",
+	  "default": "./lib/main.umd.js"
+	}
   },
   "directories": {
-    "lib": "lib"
+	"lib": "lib"
   },
   "scripts": {
-    "test": "jest --config jestconfig.json ./tests",
-    "build": "rollup -c"
+	"test": "jest --config jestconfig.json ./tests",
+	"build": "rollup -c"
   },
   "repository": {
-    "type": "git",
-    "url": "git+https://github.com/BJNSTNKVC/js-local-storage.git"
+	"type": "git",
+	"url": "git+https://github.com/BJNSTNKVC/js-local-storage.git"
   },
   "keywords": [
-    "storage"
+	"storage"
   ],
   "author": "Bojan Stankovic",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/BJNSTNKVC/js-local-storage/issues"
+	"url": "https://github.com/BJNSTNKVC/js-local-storage/issues"
   },
   "homepage": "https://github.com/BJNSTNKVC/js-local-storage#readme",
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^28.0.6",
-    "@rollup/plugin-node-resolve": "^16.0.1",
-    "@rollup/plugin-terser": "^0.4.4",
-    "@types/jest": "^29.5.14",
-    "@types/node": "^20.19.17",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "rollup": "^4.52.0",
-    "rollup-plugin-typescript2": "^0.36.0",
-    "terser": "^5.44.0",
-    "ts-jest": "^29.4.4",
-    "typescript": "^5.9.2"
+	"@rollup/plugin-commonjs": "^28.0.6",
+	"@rollup/plugin-node-resolve": "^16.0.1",
+	"@rollup/plugin-terser": "^0.4.4",
+	"@types/jest": "^29.5.14",
+	"@types/node": "^20.19.17",
+	"jest": "^29.7.0",
+	"jest-environment-jsdom": "^29.7.0",
+	"rollup": "^4.52.0",
+	"rollup-plugin-typescript2": "^0.36.0",
+	"terser": "^5.44.0",
+	"ts-jest": "^29.4.4",
+	"typescript": "^5.9.2"
   }
 }
