+++ src/storage/index.ts
@@ -1,0 +1,2 @@
+export { LocalStorage, type LocalStorageItem } from './LocalStorage';
+export { LocalStorageFake } from './LocalStorageFake';
